+++ src/main/kotlin/de/itemis/mps/gradle/runmigrations/Plugin.kt
@@ -1,6 +1,7 @@
 package de.itemis.mps.gradle.runmigrations
 
-import de.itemis.mps.gradle.*
+import de.itemis.mps.gradle.BasePluginExtensions
+import de.itemis.mps.gradle.getMPSVersion
 import net.swiftzer.semver.SemVer
 import org.gradle.api.GradleException
 import org.gradle.api.Plugin
@@ -13,11 +14,23 @@ import java.io.File
 import javax.inject.Inject
 
 open class MigrationExecutorPluginExtensions @Inject constructor(of: ObjectFactory) : BasePluginExtensions(of) {
+    /**
+     * (Since MPS 2021.1) Whether to halt if a pre-check has failed. Note that the check for migrated dependencies
+     * cannot be skipped.
+     */
+    var haltOnPrecheckFailure = true
+
+    /**
+     * (Since MPS 2021.3) Whether to force a migration even if the project directory contains `.allow-pending-migrations` file.
+     */
+
     var force = false
 }
+
 @Suppress("unused")
 open class RunMigrationsMpsProjectPlugin : Plugin<Project> {
     companion object {
+        val MIN_VERSION_FOR_HALT_ON_PRECHECK_FAILURE = SemVer(2021, 1)
         val MIN_VERSION_FOR_FORCE = SemVer(2021, 3)
     }
 
@@ -32,14 +45,18 @@ open class RunMigrationsMpsProjectPlugin : Plugin<Project> {
                 if (!file(projectLocation).exists()) {
                     throw GradleException("The path to the project doesn't exist: $projectLocation")
                 }
-                val forceMigration = extension.force
-                
+
                 val mpsVersion = extension.getMPSVersion()
                 val parsedMPSVersion = SemVer.parse(mpsVersion)
-                if (forceMigration && parsedMPSVersion < MIN_VERSION_FOR_FORCE) {
+
+                if (extension.force && parsedMPSVersion < MIN_VERSION_FOR_FORCE) {
                     throw GradleException("The force migration flag is only supported for MPS version $MIN_VERSION_FOR_FORCE and higher.")
                 }
-                
+
+                if (!extension.haltOnPrecheckFailure && parsedMPSVersion < MIN_VERSION_FOR_HALT_ON_PRECHECK_FAILURE) {
+                    throw GradleException("The 'do not halt on pre-check failure' option is only supported for MPS version $MIN_VERSION_FOR_HALT_ON_PRECHECK_FAILURE and higher.")
+                }
+
                 val resolveMps: Task = if (extension.mpsConfig != null) {
                     tasks.create("resolveMpsForMigrations", Copy::class.java) {
                         from({ extension.mpsConfig!!.resolve().map(::zipTree) })
@@ -63,16 +80,40 @@ open class RunMigrationsMpsProjectPlugin : Plugin<Project> {
                             }
                             "taskdef"("resource" to "jetbrains/mps/build/ant/antlib.xml", "classpathref" to "path.mps.ant.path")
 
-                            val baseArgsToMigrate = arrayOf("project" to projectLocation, "mpsHome" to mpsLocation)
-                            val argsToMigrate = if (forceMigration) arrayOf(*baseArgsToMigrate, "force" to true) else baseArgsToMigrate
+                            val argsToMigrate = mutableListOf<Pair<String, Any>>().run {
+                                add("project" to projectLocation)
+                                add("mpsHome" to mpsLocation)
+
+                                if (parsedMPSVersion >= MIN_VERSION_FOR_FORCE) add("force" to extension.force)
+                                if (parsedMPSVersion >= MIN_VERSION_FOR_HALT_ON_PRECHECK_FAILURE) add("haltOnPrecheckFailure" to extension.haltOnPrecheckFailure)
+
+                                toTypedArray()
+                            }
 
                             "migrate"(*argsToMigrate) {
                                 "macro"("name" to "mps_home", "path" to mpsLocation)
-                                "jvmargs"() {
+
+                                extension.macros.forEach {
+                                    "macro"("name" to it.name, "path" to it.value)
+                                }
+
+                                "jvmargs" {
                                     "arg"("value" to "-Didea.log.config.file=log.xml")
                                     "arg"("value" to "-ea")
                                     if (extension.maxHeap != null) {
                                         "arg"("value" to "-Xmx${extension.maxHeap}")
+                                    }
+                                }
+
+                                extension.pluginsProperty.get().forEach {
+                                    // Same handling as in mps-build-backends
+                                    if (File(it.path).isAbsolute) {
+                                        "plugin"("path" to it.path, "id" to it.id)
+                                    }
+                                    else if (extension.pluginLocation != null && File(extension.pluginLocation, it.path).exists()) {
+                                        "plugin"("path" to File(extension.pluginLocation, it.path), "id" to it.id)
+                                    } else {
+                                        "plugin"("path" to mpsLocation.resolve("plugins").resolve(it.path), "id" to it.id)
                                     }
                                 }
                             }